+++ app/frontend/src/javascript/models/prepaid-pack.ts
@@ -1,14 +1,15 @@
 
 export interface PackIndexFilter {
-  group_id: number,
-  priceable_id: number,
-  priceable_type: string
+  group_id?: number,
+  priceable_id?: number,
+  priceable_type?: string,
+  disabled?: boolean,
 }
 
 export interface PrepaidPack {
   id?: number,
-  priceable_id: number,
-  priceable_type: string,
+  priceable_id?: number,
+  priceable_type?: string,
   group_id: number,
   validity_interval?: 'day' | 'week' | 'month' | 'year',
   validity_count?: number,