+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog Fab-manager
 
+- [TODO DEPLOY] `rails db:seed`
+
 ## v5.0.7 2021 June 24
 
 - Fix a bug: unable to export members list if no subscriptions was taken