+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog Fab-manager
 
 - Ability to run `fablab:chain:all` non interactively
+- Full test coverage of the OpenAPI
+- Fix a bug: unable to get the Events without images from the OpenAPI
+- Fix a bug: unable to get the Space reservations from the OpenAPI
+- Fix a bug: unable to get invoices from the OpenAPI
 
 ## v5.0.7 2021 June 24